requirements.txt
requests==2.26.0
pandas==1.3.3
pandas_ta==0.3.14b0
matplotlib==3.4.3
scikit-learn==0.24.2
python-telegram-bot==13.7
python-dotenv==0.19.0
cachetools==4.2.2
ccxt==1.53.59
mplfinance==0.12.7a17
aiohttp==3.7.4.post0
